+++ packages/Webkul/Shop/src/Resources/lang/ar/app.php
@@ -1677,6 +1677,7 @@ return [
             'creation-error' => 'لا يمكن تحديث حالة RMA لأن الفاتورة الخاصة بهذا الطلب لم يتم إنشاؤها.',
             'permission-denied' => 'أنت مسجل الدخول',
             'rma-disabled' => 'تم تعطيل RMA لهذا المنتج',
+            'reopen-not-allowed' => 'غير مسموح بإعادة فتح طلب RMA هذا.',
             'send-message' => 'تم إرسال :name بنجاح.',
             'update-success' => 'تم تحديث :name بنجاح.',
         ],