+++ packages/Webkul/Shop/src/Resources/lang/de/app.php
@@ -1677,6 +1677,7 @@ return [
             'creation-error' => 'Der RMA-Status kann nicht aktualisiert werden, da die Rechnung für diese Bestellung nicht erstellt wurde.',
             'permission-denied' => 'Sie sind angemeldet',
             'rma-disabled' => 'RMA ist für dieses Produkt deaktiviert',
+            'reopen-not-allowed' => 'Das Wiedereröffnen dieser RMA-Anfrage ist nicht zulässig.',
             'send-message' => ':name erfolgreich gesendet.',
             'update-success' => ':name erfolgreich aktualisiert.',
         ],